requirements.txt
playwright
requests
pillow
openai>=1.0.0